refactor(append_receipt): split into focused modules (W1C)
Split scripts/append_receipt.py (1458L) into a thin facade + 11 focused
modules under scripts/lib/append_receipt_internals/. Pure mechanical
Extract Method — public API, CLI, and behavior unchanged.

Modules:
  - common.py             (140L) shared helpers
  - validation.py         (117L) input validation
  - idempotency.py        (169L) duplicate detection
  - git_provenance.py      (85L) git/branch/commit metadata
  - session_resolver.py   (245L) session_id, token usage, model
  - report_extractor.py    (64L) report path scanning
  - register_emit.py       (69L) dispatch register emission
  - quality.py            (154L) quality findings + open-item creation
  - enrichment.py         (274L) orchestrates enrichment phases
  - payload.py            (168L) append_receipt_payload entry

Facade (scripts/append_receipt.py): 172L (was 1458L). Re-exports
public symbols for importers. CLI argparse unchanged.

Resolves OI-1201, OI-1202, OI-1203, OI-1204, OI-1291, OI-1292, OI-1293.

Verification:
  - pytest tests/test_append_receipt*.py: 64 passed, 2 failed
    (both failures pre-exist on origin/main: missing receipt_notifier.sh)
  - 0 functions >70L (blocking threshold). 6 functions 51-57L (warn).
  - CLI --help works.
  - Importers resolve (append_receipt_payload, _enrich_completion_receipt,
    _resolve_session_id, ...).

Dispatch-ID: 20260501-w1c-append-receipt-refactor
