+++ scripts/append_receipt.py
@@ -1,19 +1,17 @@
 #!/usr/bin/env python3
-"""Canonical receipt append helper with lock, validation, and idempotency guard."""
+"""Canonical receipt append helper with lock, validation, and idempotency guard.
+
+Facade + CLI entry point. Implementation lives in
+``scripts/lib/append_receipt_internals/``.
+"""
 
 from __future__ import annotations
 
 import argparse
-import datetime as dt
-import fcntl
-import hashlib
 import json
 import os
-import re
-import subprocess
+import subprocess  # noqa: F401  (mock.patch("append_receipt.subprocess.Popen") relies on this)
 import sys
-import time
-from dataclasses import dataclass
 from pathlib import Path
 from typing import Any, Dict, List, Optional
 
@@ -32,1357 +30,73 @@ try:
 except Exception as exc:  # pragma: no cover - hard fail on bootstrap issue
     raise SystemExit(f"Failed to load vnx_paths: {exc}")
 
-# Lazy import for open_items_manager (only needed for quality registration)
-_open_items_manager = None
-
-def _get_open_items_manager():
-    global _open_items_manager
-    if _open_items_manager is None:
-        sys.path.insert(0, str(SCRIPT_DIR))
-        import open_items_manager as _oim
-        _open_items_manager = _oim
-    return _open_items_manager
-
-EXIT_OK = 0
-EXIT_INVALID_INPUT = 10
-EXIT_VALIDATION_ERROR = 11
-EXIT_IO_ERROR = 12
-EXIT_LOCK_ERROR = 13
-EXIT_UNEXPECTED_ERROR = 20
-
-DISPATCH_REQUIRED_EVENTS = {
-    "task_started",
-    "task_complete",
-    "task_failed",
-    "task_timeout",
-    "task_blocked",
-    "dispatch_sent",
-    "dispatch_ack",
-    "ack",
-}
-
-STATE_MUTATION_EVENTS = {"state_mutation"}
-
-_REPO_ROOT = SCRIPT_DIR.parent
-_REBUILD_THROTTLE_SECONDS = 30
-
-_warned_review_gate_no_dispatch_id = False
-
-
-IDEMPOTENCY_FIELDS = (
-    "dispatch_id",
-    "task_id",
-    "pr_number",  # prevents review_gate_request fan-out collision per gate
-    "gate",       # multiple gates per dispatch_id must not collide
-    "terminal",
-    "event_type",
-    "event",
-    "report_path",
-    "source",
-    "file",
-    "trigger",
-    "section",
+from append_receipt_internals.common import (
+    AppendReceiptError,
+    AppendResult,
+    EXIT_INVALID_INPUT,
+    EXIT_IO_ERROR,
+    EXIT_OK,
+    EXIT_UNEXPECTED_ERROR,
+    REPO_ROOT as _REPO_ROOT,
+    _emit,
+    _get_open_items_manager,
+    _safe_subprocess,
+    _utc_now_iso,
+    is_headless_t0,
+    register_facade,
+)
+from append_receipt_internals.idempotency import (
+    IDEMPOTENCY_FIELDS,
+    _cache_file_for,
+    _compute_idempotency_key,
+    _load_cache,
+    _lock_file_for,
+    _resolve_receipts_file,
+    _write_cache,
+    _write_receipt_under_lock,
+)
+from append_receipt_internals.validation import (
+    DISPATCH_REQUIRED_EVENTS,
+    STATE_MUTATION_EVENTS,
+    _is_completion_event,
+    _is_subprocess_intermediate_completion,
+    _maybe_reroute_ghost_receipt,
+    _requires_dispatch_id,
+    _validate_receipt,
+    _warn_if_review_gate_missing_dispatch_id,
+)
+from append_receipt_internals.report_extractor import _extract_changed_files_from_report
+from append_receipt_internals.git_provenance import (
+    _build_git_provenance,
+    _extract_shortstat_value,
+)
+from append_receipt_internals.session_resolver import (
+    _build_session_metadata,
+    _extract_session_token_usage,
+    _resolve_model_provider,
+    _resolve_session_id,
+    _rsi_check_env_session,
+    _rsi_check_provider_files,
+)
+from append_receipt_internals.quality import (
+    _SEVERITY_MAP,
+    _count_quality_violations,
+    _count_quality_violations_against_store,
+    _register_quality_open_items,
+)
+from append_receipt_internals.register_emit import _emit_dispatch_register
+from append_receipt_internals.enrichment import _enrich_completion_receipt
+from append_receipt_internals.payload import (
+    _maybe_trigger_state_rebuild,
+    _run_post_append_hooks,
+    _trigger_receipt_classifier,
+    _update_confidence_from_receipt,
+    append_receipt_payload,
 )
 
-
-def is_headless_t0() -> bool:
-    """Return True when T0 is configured to run via subprocess adapter.
-
-    Set VNX_ADAPTER_T0=subprocess to enable headless T0 mode.  When True,
-    the receipt processor annotates the T0 terminal snapshot entry with
-    adapter/headless metadata and skips tmux-dependent probes for T0.
-    """
-    return os.environ.get("VNX_ADAPTER_T0", "tmux").lower() == "subprocess"
-
-
-@dataclass(frozen=True)
-class AppendResult:
-    status: str
-    receipts_file: Path
-    idempotency_key: str
-
-
-class AppendReceiptError(RuntimeError):
-    def __init__(self, code: str, exit_code: int, message: str):
-        super().__init__(message)
-        self.code = code
-        self.exit_code = exit_code
-        self.message = message
-
-
-def _emit(level: str, code: str, **fields: Any) -> None:
-    payload = {
-        "level": level,
-        "code": code,
-        "timestamp": int(time.time()),
-    }
-    payload.update(fields)
-    print(json.dumps(payload, separators=(",", ":"), sort_keys=True), file=sys.stderr)
-
-
-def _resolve_receipts_file(receipts_file: Optional[str] = None) -> Path:
-    if receipts_file:
-        return Path(receipts_file).expanduser()
-
-    paths = ensure_env()
-    return Path(paths["VNX_STATE_DIR"]) / "t0_receipts.ndjson"
-
-
-def _lock_file_for(receipts_path: Path) -> Path:
-    return receipts_path.parent / "append_receipt.lock"
-
-
-def _cache_file_for(receipts_path: Path) -> Path:
-    return receipts_path.parent / "receipt_idempotency_recent.ndjson"
-
-
-def _requires_dispatch_id(receipt: Dict[str, Any], event_name: str) -> bool:
-    if event_name in DISPATCH_REQUIRED_EVENTS:
-        return True
-    if event_name.startswith("task_"):
-        return True
-    if receipt.get("task_id"):
-        return True
-    return False
-
-
-def _warn_if_review_gate_missing_dispatch_id(event_name: str, receipt: Dict[str, Any]) -> None:
-    global _warned_review_gate_no_dispatch_id
-    if _warned_review_gate_no_dispatch_id:
-        return
-    if event_name == "review_gate_request":
-        if not str(receipt.get("dispatch_id", "")).strip():
-            _warned_review_gate_no_dispatch_id = True
-            _emit(
-                "WARN",
-                "review_gate_request_missing_dispatch_id",
-                message="review_gate_request receipt has no dispatch_id — receipt-to-gate audit linkage severed",
-            )
-
-
-def _validate_receipt(receipt: Dict[str, Any]) -> str:
-    timestamp = str(receipt.get("timestamp", "")).strip()
-    if not timestamp:
-        raise AppendReceiptError(
-            "missing_required_key",
-            EXIT_VALIDATION_ERROR,
-            "Missing required key: timestamp",
-        )
-
-    event_name = str(receipt.get("event_type") or receipt.get("event") or "").strip()
-    if not event_name:
-        raise AppendReceiptError(
-            "missing_required_key",
-            EXIT_VALIDATION_ERROR,
-            "Missing required key: event_type or event",
-        )
-
-    if _requires_dispatch_id(receipt, event_name):
-        dispatch_id = str(receipt.get("dispatch_id", "")).strip()
-        if not dispatch_id:
-            raise AppendReceiptError(
-                "missing_required_key",
-                EXIT_VALIDATION_ERROR,
-                "Missing required key: dispatch_id",
-            )
-
-    _warn_if_review_gate_missing_dispatch_id(event_name, receipt)
-
-    return event_name
-
-
-def _compute_idempotency_key(receipt: Dict[str, Any], event_name: str) -> str:
-    digest_fields: Dict[str, Any] = {}
-
-    for field in IDEMPOTENCY_FIELDS:
-        value = receipt.get(field)
-        if value is None:
-            continue
-        if isinstance(value, str) and not value.strip():
-            continue
-        digest_fields[field] = value
-
-    if "event_type" not in digest_fields and "event" not in digest_fields:
-        digest_fields["event_type"] = event_name
-
-    # For receipts without stable identity fields, include timestamp to avoid
-    # collapsing distinct events in the short dedupe window.
-    if (
-        "dispatch_id" not in digest_fields
-        and "task_id" not in digest_fields
-        and "report_path" not in digest_fields
-    ):
-        digest_fields["timestamp"] = receipt.get("timestamp")
-
-    if not digest_fields:
-        digest_fields = receipt
-
-    payload = json.dumps(digest_fields, sort_keys=True, separators=(",", ":"))
-    return hashlib.sha256(payload.encode("utf-8")).hexdigest()
-
-
-def _load_cache(cache_file: Path, min_epoch: float) -> List[Dict[str, Any]]:
-    if not cache_file.exists():
-        return []
-
-    entries: List[Dict[str, Any]] = []
-    try:
-        with cache_file.open("r", encoding="utf-8") as handle:
-            for raw_line in handle:
-                line = raw_line.strip()
-                if not line:
-                    continue
-                try:
-                    parsed = json.loads(line)
-                except json.JSONDecodeError:
-                    continue
-                ts = float(parsed.get("ts", 0))
-                key = str(parsed.get("key", "")).strip()
-                if key and ts >= min_epoch:
-                    entries.append({"ts": ts, "key": key})
-    except OSError as exc:
-        raise AppendReceiptError("cache_read_failed", EXIT_IO_ERROR, f"Failed to read idempotency cache: {exc}") from exc
-
-    return entries
-
-
-def _write_cache(cache_file: Path, entries: List[Dict[str, Any]], max_entries: int = 2048) -> None:
-    entries = entries[-max_entries:]
-    tmp_file = cache_file.with_name(f"{cache_file.name}.{os.getpid()}.tmp")
-
-    try:
-        with tmp_file.open("w", encoding="utf-8") as handle:
-            for entry in entries:
-                handle.write(json.dumps(entry, sort_keys=True, separators=(",", ":")))
-                handle.write("\n")
-        os.replace(tmp_file, cache_file)
-    except OSError as exc:
-        raise AppendReceiptError("cache_write_failed", EXIT_IO_ERROR, f"Failed to write idempotency cache: {exc}") from exc
-    finally:
-        try:
-            if tmp_file.exists():
-                tmp_file.unlink()
-        except OSError:
-            pass
-
-
-def _is_completion_event(receipt: Dict[str, Any]) -> bool:
-    """Check if receipt is a completion event."""
-    event_type = receipt.get("event_type") or receipt.get("event") or ""
-    return event_type in (
-        "task_complete",
-        "task_completed",
-        "completion",
-        "complete",
-        "subprocess_completion",
-    )
-
-
-def _is_subprocess_intermediate_completion(receipt: Dict[str, Any]) -> bool:
-    """True for the intermediate subprocess-adapter completion receipt.
-
-    These receipts are appended when the subprocess exits but BEFORE the
-    real report has been extracted (subprocess_adapter only drops an async
-    trigger file at that point). They typically lack ``report_path`` and a
-    git diff against HEAD will report no changed files, so generating a
-    quality advisory or persisting CQS would overwrite ``dispatch_metadata``
-    with synthetic "No changed files detected" data and corrupt the row
-    permanently if the downstream report-driven enrichment is delayed or
-    fails.
-
-    The session/provenance/snapshot enrichment is still safe and desirable
-    for these receipts (e.g. instruction_sha256 surfacing).
-    """
-    event_type = receipt.get("event_type") or receipt.get("event") or ""
-    return event_type == "subprocess_completion"
-
-
-def _safe_subprocess(cmd: List[str], cwd: Optional[Path] = None, timeout: int = 5) -> Optional[str]:
-    try:
-        result = subprocess.run(
-            cmd,
-            cwd=str(cwd) if cwd else None,
-            capture_output=True,
-            text=True,
-            check=False,
-            timeout=timeout,
-        )
-    except Exception:
-        return None
-    if result.returncode != 0:
-        return None
-    return (result.stdout or "").strip()
-
-
-def _utc_now_iso() -> str:
-    return dt.datetime.now(dt.timezone.utc).replace(microsecond=0).isoformat().replace("+00:00", "Z")
-
-
-def _extract_shortstat_value(shortstat: str, token: str) -> int:
-    # Example: "12 files changed, 342 insertions(+), 87 deletions(-)"
-    for part in shortstat.split(","):
-        chunk = part.strip().lower()
-        if token in chunk:
-            digits = "".join(ch for ch in chunk if ch.isdigit())
-            if digits:
-                try:
-                    return int(digits)
-                except ValueError:
-                    return 0
-    return 0
-
-
-def _build_git_provenance(repo_root: Path) -> Dict[str, Any]:
-    # Resolve the PROJECT root, not the vnx-system root.
-    # CLAUDE_PROJECT_DIR points to the actual project worktree.
-    project_dir = os.environ.get("CLAUDE_PROJECT_DIR", "")
-    if project_dir:
-        probe = _safe_subprocess(["git", "rev-parse", "--show-toplevel"], cwd=Path(project_dir))
-        if probe:
-            repo_root = Path(project_dir)
-
-    git_root_raw = _safe_subprocess(["git", "rev-parse", "--show-toplevel"], cwd=repo_root)
-    captured_at = _utc_now_iso()
-
-    if not git_root_raw:
-        return {
-            "git_ref": "not_a_repo",
-            "branch": "unknown",
-            "is_dirty": False,
-            "dirty_files": 0,
-            "diff_summary": None,
-            "captured_at": captured_at,
-            "captured_by": "append_receipt",
-        }
-
-    git_root = Path(git_root_raw)
-    git_ref = _safe_subprocess(["git", "rev-parse", "HEAD"], cwd=git_root) or "unknown"
-    branch = _safe_subprocess(["git", "rev-parse", "--abbrev-ref", "HEAD"], cwd=git_root) or "unknown"
-    status_raw = _safe_subprocess(["git", "status", "--porcelain"], cwd=git_root) or ""
-    dirty_files = len([line for line in status_raw.splitlines() if line.strip()])
-    is_dirty = dirty_files > 0
-
-    diff_summary = None
-    if is_dirty:
-        shortstat = _safe_subprocess(["git", "diff", "--shortstat"], cwd=git_root) or ""
-        if shortstat:
-            diff_summary = {
-                "files_changed": _extract_shortstat_value(shortstat, "file"),
-                "insertions": _extract_shortstat_value(shortstat, "insertion"),
-                "deletions": _extract_shortstat_value(shortstat, "deletion"),
-            }
-
-    # Detect if running inside a git worktree
-    git_dir = _safe_subprocess(["git", "rev-parse", "--git-dir"], cwd=git_root) or ""
-    git_common_dir = _safe_subprocess(["git", "rev-parse", "--git-common-dir"], cwd=git_root) or ""
-    in_worktree = bool(git_dir and git_common_dir and git_dir != git_common_dir)
-
-    provenance = {
-        "git_ref": git_ref,
-        "branch": branch,
-        "is_dirty": is_dirty,
-        "dirty_files": dirty_files,
-        "diff_summary": diff_summary,
-        "in_worktree": in_worktree,
-        "captured_at": captured_at,
-        "captured_by": "append_receipt",
-    }
-    if in_worktree:
-        provenance["worktree_path"] = str(git_root)
-    return provenance
-
-
-_PROVIDER_ENV_VAR: Dict[str, str] = {
-    "claude_code": "CLAUDE_SESSION_ID",
-    "gemini_cli": "GEMINI_SESSION_ID",
-    "codex_cli": "CODEX_SESSION_ID",
-    "kimi_cli": "KIMI_SESSION_ID",
-}
-
-_PROVIDER_SESSION_FILES: Dict[str, Path] = {
-    "claude_code": Path.home() / ".claude" / "sessions" / "current",
-    "gemini_cli": Path.home() / ".gemini" / "sessions" / "current",
-    "codex_cli": Path.home() / ".codex" / "sessions" / "current",
-    "kimi_cli": Path.home() / ".kimi" / "sessions" / "current",
-}
-
-
-def _rsi_check_env_session(
-    terminal: str, state_dir: Path, current_session_file: Path
-) -> Optional[str]:
-    """Priority 3: check provider env var; auto-create per-terminal file if found."""
-    mp = _resolve_model_provider(terminal, state_dir)
-    provider = mp.get("provider", "unknown")
-    env_var = _PROVIDER_ENV_VAR.get(provider, "CLAUDE_SESSION_ID")
-    env_value = os.environ.get(env_var)
-    if env_value:
-        value = env_value.strip()
-        if value and value not in {"unknown", "null", "None"}:
-            try:
-                state_dir.mkdir(parents=True, exist_ok=True)
-                current_session_file.write_text(value, encoding="utf-8")
-            except Exception:
-                pass
-            return value
-    return None
-
-
-def _rsi_check_provider_files(terminal: str, state_dir: Path) -> Optional[str]:
-    """Priority 4: check provider session files in provider-priority order."""
-    mp = _resolve_model_provider(terminal, state_dir)
-    provider = mp.get("provider", "unknown")
-    preferred_file = _PROVIDER_SESSION_FILES.get(provider)
-    other_files = [f for f in _PROVIDER_SESSION_FILES.values() if f != preferred_file]
-    files = ([preferred_file] + other_files) if preferred_file else list(_PROVIDER_SESSION_FILES.values())
-    for current_file in files:
-        try:
-            if current_file.exists():
-                value = current_file.read_text(encoding="utf-8").strip()
-                if value and value not in {"unknown", "null", "None"}:
-                    return value
-        except Exception:
-            continue
-    return None
-
-
-def _resolve_session_id(receipt: Dict[str, Any], state_dir: Optional[Path] = None) -> str:
-    """Resolve session_id with deterministic priority chain (parallel-terminal safe).
-
-    Priority chain (matches session_resolver.sh):
-    1. Report-provided session_id (explicit in metadata)
-    2. Per-terminal current_session files (deterministic, parallel-safe)
-    3. Environment variables (with auto-create of per-terminal files)
-    4. Provider "current" files (global session files)
-    5. Fallback: "unknown"
-
-    Args:
-        receipt: Receipt payload.
-        state_dir: Resolved state directory (caller should supply to avoid
-                   double-resolution). Falls back to resolve_state_dir(__file__).
-    """
-    metadata = receipt.get("metadata") if isinstance(receipt.get("metadata"), dict) else {}
-    terminal = str(receipt.get("terminal") or "unknown").strip()
-
-    # Priority 1: Report-provided session_id (explicit - RECOMMENDED)
-    for candidate in (metadata.get("session_id"), metadata.get("session"), receipt.get("session_id")):
-        value = str(candidate or "").strip()
-        if value and value not in {"unknown", "null", "None"}:
-            return value
-
-    # Priority 2: Per-terminal current_session file (DETERMINISTIC)
-    if state_dir is None:
-        state_dir = resolve_state_dir(__file__)
-    current_session_file = state_dir / f"current_session_{terminal}"
-    if current_session_file.exists():
-        try:
-            value = current_session_file.read_text(encoding="utf-8").strip()
-            if value and value not in {"unknown", "null", "None"}:
-                return value
-        except Exception:
-            pass
-
-    # Priority 3: Environment variables (provider-specific) with auto-create
-    value = _rsi_check_env_session(terminal, state_dir, current_session_file)
-    if value:
-        return value
-
-    # Priority 4: Provider "current" session files (preferred provider first)
-    value = _rsi_check_provider_files(terminal, state_dir)
-    if value:
-        return value
-
-    return "unknown"
-
-
-def _resolve_model_provider(terminal: str, state_dir: Path) -> Dict[str, str]:
-    """Resolve model and provider with panes.json priority (matches session_resolver.sh).
-
-    Priority:
-    1. panes.json mapping (if exists)
-    2. Terminal naming convention heuristic (fallback)
-
-    Default models for Claude terminals:
-    - T0: claude-opus-4.6
-    - T1/T2/T3/T-MANAGER: claude-sonnet-4.5
-    """
-    model = "unknown"
-    provider = "unknown"
-
-    # Priority 1: panes.json mapping (if exists)
-    panes_json = state_dir / "panes.json"
-    if panes_json.exists():
-        try:
-            payload = json.loads(panes_json.read_text(encoding="utf-8"))
-            if isinstance(payload, dict):
-                term_lower = terminal.lower()
-                entry = payload.get(terminal) or payload.get(term_lower)
-                if isinstance(entry, dict):
-                    model = str(entry.get("model") or "unknown").strip() or "unknown"
-                    provider = str(entry.get("provider") or "unknown").strip().lower() or "unknown"
-        except Exception:
-            pass
-
-    # Priority 2: Terminal naming convention heuristic (fallback)
-    # Model is NOT guessed — it must come from panes.json or remain "unknown".
-    # Hardcoding model defaults leads to incorrect receipt metadata when
-    # terminals run a different model than the startup default.
-    if provider == "unknown":
-        upper = terminal.upper()
-        if upper in ("T0", "T1", "T2", "T3", "T-MANAGER"):
-            provider = "claude_code"
-        elif "GEMINI" in upper or upper.startswith("GEM-"):
-            provider = "gemini_cli"
-            if model == "unknown":
-                model = "gemini-pro"
-        elif "CODEX" in upper or upper.startswith("CODE-"):
-            provider = "codex_cli"
-            if model == "unknown":
-                model = "gpt-5.2-codex"
-        elif "KIMI" in upper:
-            provider = "kimi_cli"
-
-    return {"model": model, "provider": provider}
-
-
-def _extract_session_token_usage(session_id: str, terminal: str) -> Optional[Dict[str, int]]:
-    """Extract cumulative token usage from Claude Code JSONL session log.
-
-    Scans the session JSONL for message.usage fields and sums all token counters.
-    Returns None if the session file cannot be found or parsed.
-    """
-    if not session_id or session_id == "unknown":
-        return None
-
-    # Claude Code stores sessions under ~/.claude/projects/<project-key>/<session-id>.jsonl
-    claude_projects = Path.home() / ".claude" / "projects"
-    if not claude_projects.is_dir():
-        return None
-
-    # Search for the session JSONL across all project dirs
-    session_file = None
-    for project_dir in claude_projects.iterdir():
-        if not project_dir.is_dir():
-            continue
-        candidate = project_dir / f"{session_id}.jsonl"
-        if candidate.is_file():
-            session_file = candidate
-            break
-
-    if not session_file:
-        return None
-
-    totals: Dict[str, int] = {
-        "input_tokens": 0,
-        "output_tokens": 0,
-        "cache_read_input_tokens": 0,
-        "cache_creation_input_tokens": 0,
-    }
-    try:
-        with open(session_file, "r", encoding="utf-8") as fh:
-            for line in fh:
-                line = line.strip()
-                if not line:
-                    continue
-                try:
-                    msg = json.loads(line)
-                    usage = (msg.get("message") or {}).get("usage")
-                    if isinstance(usage, dict):
-                        for key in totals:
-                            totals[key] += usage.get(key, 0)
-                except (json.JSONDecodeError, AttributeError):
-                    continue
-    except (OSError, IOError):
-        return None
-
-    # Only return if we actually found token data
-    if totals["input_tokens"] == 0 and totals["output_tokens"] == 0:
-        return None
-    return totals
-
-
-def _build_session_metadata(receipt: Dict[str, Any], state_dir: Path) -> Dict[str, Any]:
-    terminal = str(receipt.get("terminal") or "unknown").strip() or "unknown"
-    model_provider = _resolve_model_provider(terminal, state_dir)
-    session_id = _resolve_session_id(receipt, state_dir)
-    metadata: Dict[str, Any] = {
-        "session_id": session_id,
-        "terminal": terminal,
-        "model": model_provider["model"],
-        "provider": model_provider["provider"],
-        "captured_at": _utc_now_iso(),
-    }
-
-    # Inject token usage — provider-specific, best-effort
-    provider = model_provider["provider"]
-    try:
-        if provider == "codex_cli":
-            from adapters.codex_adapter import CodexAdapter  # noqa: PLC0415
-            token_usage = CodexAdapter.get_token_usage(terminal, state_dir)
-        elif provider in ("gemini_cli", "gemini"):
-            from adapters.gemini_adapter import GeminiAdapter  # noqa: PLC0415
-            token_usage = GeminiAdapter.get_token_usage(terminal, state_dir)
-        else:
-            # Claude (and other) terminals: scan session JSONL (existing path)
-            token_usage = _extract_session_token_usage(session_id, terminal)
-        if token_usage:
-            metadata["token_usage"] = token_usage
-    except Exception:
-        pass
-
-    # Inject instruction_sha256 from manifest (best-effort)
-    manifest_path = receipt.get("manifest_path")
-    if manifest_path:
-        try:
-            manifest_data = json.loads(Path(manifest_path).read_text())
-            sha = manifest_data.get("instruction_sha256")
-            if sha:
-                metadata["instruction_sha256"] = sha
-        except (OSError, IOError, json.JSONDecodeError) as exc:
-            _emit("WARN", "manifest_sha256_read_failed", manifest_path=str(manifest_path), error=str(exc))
-
-    return metadata
-
-
-def _enrich_completion_receipt(receipt: Dict[str, Any], repo_root: Optional[Path] = None) -> Dict[str, Any]:
-    """Enrich completion receipts with quality advisory and terminal snapshot.
-
-    This is best-effort - failures will result in status="unavailable" markers
-    rather than crashing the receipt append flow.
-
-    Args:
-        receipt: Receipt payload to enrich
-        repo_root: Repository root path for git operations
-
-    Returns:
-        Enriched receipt with quality_advisory and terminal_snapshot fields
-    """
-    # Only enrich completion receipts
-    if not _is_completion_event(receipt):
-        return receipt
-
-    enriched = receipt.copy()
-    paths = ensure_env()
-    state_dir = Path(paths.get("VNX_STATE_DIR", ".")).resolve()
-
-    # Stamp project_id on the receipt for cross-DB attribution.
-    # current_project_id() falls back to 'vnx-dev' so vnx-dev call sites are
-    # unchanged from Phase 0.
-    enriched.setdefault("project_id", current_project_id())
-
-    # Inject git provenance metadata (best-effort).
-    if "provenance" not in enriched:
-        try:
-            resolved_repo_root = repo_root or Path(paths.get("PROJECT_ROOT", Path.cwd())).resolve()
-            enriched["provenance"] = _build_git_provenance(resolved_repo_root)
-        except Exception as exc:
-            _emit("WARN", "provenance_capture_failed", error=str(exc))
-            enriched["provenance"] = {
-                "git_ref": "unknown",
-                "branch": "unknown",
-                "is_dirty": False,
-                "dirty_files": 0,
-                "diff_summary": None,
-                "captured_at": _utc_now_iso(),
-                "captured_by": "append_receipt",
-                "status": "unavailable",
-                "error": str(exc),
-            }
-
-    # Inject session metadata for usage correlation (best-effort).
-    existing_session = enriched.get("session")
-    if isinstance(existing_session, dict):
-        merged_session = dict(existing_session)
-        try:
-            defaults = _build_session_metadata(enriched, state_dir)
-            for key, value in defaults.items():
-                merged_session.setdefault(key, value)
-        except Exception as exc:
-            _emit("WARN", "session_metadata_failed", error=str(exc))
-        enriched["session"] = merged_session
-    else:
-        try:
-            enriched["session"] = _build_session_metadata(enriched, state_dir)
-        except Exception as exc:
-            _emit("WARN", "session_metadata_failed", error=str(exc))
-            enriched["session"] = {
-                "session_id": "unknown",
-                "terminal": str(enriched.get("terminal") or "unknown"),
-                "model": "unknown",
-                "provider": "unknown",
-                "captured_at": _utc_now_iso(),
-                "status": "unavailable",
-                "error": str(exc),
-            }
-
-    # Enrich with provenance linkage fields (PR-2: dispatch_id, trace_token, etc.)
-    try:
-        enrich_receipt_provenance(enriched)
-        # Validate provenance and log gaps (best-effort)
-        prov_validation = validate_receipt_provenance(enriched)
-        if prov_validation.gaps:
-            gap_summaries = [g.to_dict() for g in prov_validation.gaps]
-            enriched.setdefault("provenance_validation", {
-                "chain_status": prov_validation.chain_status,
-                "gaps": gap_summaries,
-            })
-            for gap in prov_validation.gaps:
-                if gap.severity in ("warning", "error"):
-                    _emit("WARN", "provenance_gap_detected",
-                           gap_type=gap.gap_type, entity_id=gap.entity_id,
-                           description=gap.description)
-    except Exception as exc:
-        _emit("WARN", "provenance_enrichment_failed", error=str(exc))
-
-    # Collect terminal snapshot (best-effort)
-    try:
-        snapshot = collect_terminal_snapshot(state_dir)
-        enriched["terminal_snapshot"] = snapshot.to_dict()
-    except Exception as exc:
-        _emit("WARN", "terminal_snapshot_failed", error=str(exc))
-        enriched["terminal_snapshot"] = {
-            "status": "unavailable",
-            "error": str(exc),
-        }
-
-    # Annotate T0 terminal snapshot entry when T0 runs via subprocess adapter.
-    # Skips tmux-based probes for T0 and marks the entry so downstream readers
-    # know T0 was headless at receipt time.
-    if is_headless_t0():
-        snapshot_data = enriched.get("terminal_snapshot") or {}
-        terminals = snapshot_data.get("terminals") or {}
-        t0_entry = dict(terminals.get("T0") or {})
-        t0_entry["adapter"] = "subprocess"
-        t0_entry["headless"] = True
-        terminals["T0"] = t0_entry
-        snapshot_data["terminals"] = terminals
-        enriched["terminal_snapshot"] = snapshot_data
-
-    # Subprocess intermediate completions skip quality-advisory generation and
-    # CQS persistence entirely. The real report is not yet extracted at this
-    # point (subprocess_adapter only drops an async trigger file), so a git
-    # diff would yield zero changed files and we'd write a synthetic
-    # "No changed files detected" advisory + zero-CQS row that overwrites
-    # any later, report-driven enrichment for this dispatch.
-    if _is_subprocess_intermediate_completion(receipt):
-        return enriched
-
-    # Generate quality advisory (best-effort)
-    try:
-        if repo_root is None:
-            # Try to detect repo root from receipt context or use current directory
-            repo_root = Path.cwd()
-
-        changed_files = get_changed_files(repo_root)
-
-        # Fallback: parse report for "Files Modified" when git diff is empty.
-        if not changed_files:
-            report_path = str(receipt.get("report_path") or "")
-            if report_path:
-                changed_files = _extract_changed_files_from_report(Path(report_path), repo_root)
-
-        # Only run quality checks if there are changed files
-        if changed_files:
-            advisory = generate_quality_advisory(changed_files, repo_root)
-            enriched["quality_advisory"] = advisory.to_dict()
-        else:
-            # No changed files - generate minimal advisory
-            enriched["quality_advisory"] = {
-                "version": "1.0",
-                "generated_at": time.strftime("%Y-%m-%dT%H:%M:%SZ", time.gmtime()),
-                "scope": [],
-                "checks": [],
-                "summary": {
-                    "warning_count": 0,
-                    "blocking_count": 0,
-                    "risk_score": 0,
-                },
-                "t0_recommendation": {
-                    "decision": "approve",
-                    "reason": "No changed files detected",
-                    "suggested_dispatches": [],
-                    "open_items": [],
-                },
-            }
-    except Exception as exc:
-        _emit("WARN", "quality_advisory_failed", error=str(exc))
-        enriched["quality_advisory"] = {
-            "status": "unavailable",
-            "error": str(exc),
-        }
-
-    # Enrich with open items delta for CQS (best-effort)
-    try:
-        dispatch_id_for_oi = enriched.get("dispatch_id") or enriched.get("metadata", {}).get("dispatch_id")
-        if dispatch_id_for_oi:
-            oim = _get_open_items_manager()
-            resolved_count = oim.count_items_closed_by_dispatch(dispatch_id_for_oi)
-            enriched["open_items_resolved"] = resolved_count
-
-            # Load target_open_items from dispatch_metadata
-            db_path_oi = state_dir / "quality_intelligence.db"
-            if db_path_oi.exists():
-                import sqlite3 as _sqlite3_oi
-                conn_oi = _sqlite3_oi.connect(str(db_path_oi))
-                try:
-                    cols = conn_oi.execute(
-                        "PRAGMA table_info(dispatch_metadata)"
-                    ).fetchall()
-                    has_project_oi = any(c[1] == "project_id" for c in cols)
-                except _sqlite3_oi.Error:
-                    has_project_oi = False
-                if has_project_oi:
-                    row = conn_oi.execute(
-                        "SELECT target_open_items FROM dispatch_metadata "
-                        "WHERE dispatch_id=? AND project_id=?",
-                        (dispatch_id_for_oi, current_project_id()),
-                    ).fetchone()
-                else:
-                    row = conn_oi.execute(
-                        "SELECT target_open_items FROM dispatch_metadata WHERE dispatch_id=?",
-                        (dispatch_id_for_oi,),
-                    ).fetchone()
-                conn_oi.close()
-                if row and row[0]:
-                    try:
-                        enriched["target_open_items"] = json.loads(row[0])
-                    except (json.JSONDecodeError, TypeError):
-                        pass
-    except Exception as exc:
-        _emit("WARN", "oi_delta_enrichment_failed", error=str(exc))
-
-    # Compute the would-be creation count without mutating the open-items store.
-    # The actual _register_quality_open_items() call is deferred until after the
-    # idempotency check in append_receipt_payload(), so a replayed receipt skipped
-    # as a duplicate cannot still mutate open_items.json or CQS state.
-    if "open_items_created" not in enriched:
-        enriched["open_items_created"] = _count_quality_violations_against_store(enriched)
-
-    # Compute CQS and persist to dispatch_metadata (best-effort).
-    # quality_advisory_json is stored so update_dispatch_cqs.py can round-trip it
-    # without losing T0 advisory and OI-delta scoring (OI-1175).
-    try:
-        db_path = state_dir / "quality_intelligence.db"
-        if db_path.exists():
-            dispatch_id = enriched.get("dispatch_id") or enriched.get("metadata", {}).get("dispatch_id")
-            if dispatch_id:
-                cqs_result = calculate_cqs(enriched, None, db_path, dispatch_id)
-                enriched["cqs"] = cqs_result
-                import sqlite3
-                conn = sqlite3.connect(str(db_path))
-                # Scope by project_id when the column exists so cross-tenant
-                # dispatch_id collisions cannot stomp another tenant's CQS.
-                has_project = False
-                try:
-                    rows = conn.execute(
-                        "PRAGMA table_info(dispatch_metadata)"
-                    ).fetchall()
-                    has_project = any(r[1] == "project_id" for r in rows)
-                except sqlite3.Error:
-                    has_project = False
-                if has_project:
-                    conn.execute(
-                        """UPDATE dispatch_metadata
-                           SET cqs=?, normalized_status=?, cqs_components=?,
-                               open_items_created=?, open_items_resolved=?,
-                               quality_advisory_json=?
-                           WHERE dispatch_id=? AND project_id=?""",
-                        (
-                            cqs_result["cqs"],
-                            cqs_result["normalized_status"],
-                            json.dumps(cqs_result["components"]),
-                            enriched.get("open_items_created", 0),
-                            enriched.get("open_items_resolved", 0),
-                            json.dumps(enriched.get("quality_advisory") or {}),
-                            dispatch_id,
-                            current_project_id(),
-                        ),
-                    )
-                else:
-                    conn.execute(
-                        """UPDATE dispatch_metadata
-                           SET cqs=?, normalized_status=?, cqs_components=?,
-                               open_items_created=?, open_items_resolved=?,
-                               quality_advisory_json=?
-                           WHERE dispatch_id=?""",
-                        (
-                            cqs_result["cqs"],
-                            cqs_result["normalized_status"],
-                            json.dumps(cqs_result["components"]),
-                            enriched.get("open_items_created", 0),
-                            enriched.get("open_items_resolved", 0),
-                            json.dumps(enriched.get("quality_advisory") or {}),
-                            dispatch_id,
-                        ),
-                    )
-                conn.commit()
-                conn.close()
-    except Exception as exc:
-        _emit("WARN", "cqs_calculation_failed", error=str(exc))
-
-    return enriched
-
-
-def _extract_changed_files_from_report(report_path: Path, repo_root: Path) -> List[Path]:
-    """Best-effort: parse 'Files Modified' section from report markdown.
-
-    Supports two formats:
-    1. Bullet list:  - `path/to/file.py` — description
-    2. Markdown table:  | `path/to/file.py` | Type | Description |
-    """
-    if not report_path.exists():
-        return []
-
-    try:
-        content = report_path.read_text(encoding="utf-8", errors="ignore")
-    except OSError:
-        return []
-
-    # Find the "Files Modified" section (## or ### heading).
-    pattern = re.compile(
-        r"^#{2,3}\s+Files\s+Modified(?:/Created)?\s*$", re.MULTILINE
-    )
-    match = pattern.search(content)
-    if not match:
-        return []
-
-    section = content[match.end():]
-    next_heading = re.search(r"^##+\s+", section, re.MULTILINE)
-    if next_heading:
-        section = section[:next_heading.start()]
-
-    files: List[Path] = []
-    for line in section.splitlines():
-        line = line.strip()
-
-        # Skip table header separators (|---|---|)
-        if re.match(r"^\|[\s\-:|]+\|$", line):
-            continue
-
-        # Extract backtick paths from any format (bullets, tables, plain)
-        backtick = re.search(r"`([^`]+\.\w+)`", line)
-        if backtick:
-            raw_path = backtick.group(1).strip()
-        elif line.startswith("-"):
-            # Fallback: "- path/to/file.py: description"
-            raw_path = line.lstrip("-").strip().split(":", 1)[0].strip()
-        elif line.startswith("|"):
-            # Table row without backticks: | path/to/file.py | ...
-            cells = [c.strip() for c in line.split("|") if c.strip()]
-            raw_path = cells[0] if cells else ""
-        else:
-            continue
-
-        if not raw_path or not re.search(r"\.\w+$", raw_path):
-            continue
-
-        candidate = Path(raw_path)
-        if not candidate.is_absolute():
-            candidate = (repo_root / candidate).resolve()
-        if candidate.exists():
-            files.append(candidate)
-
-    return files
-
-
-_SEVERITY_MAP = {
-    "blocking": "blocker",
-    "warning": "warn",
-    "info": "info",
-}
-
-
-def _emit_dispatch_register(receipt: dict) -> bool:
-    """Emit dispatch_register event for codex_gate-relevant receipts.
-
-    SCOPE: codex_gate only. gemini_review and claude_github_optional are
-    deferred until proper findings parsers exist (separate PR).
-
-    Returns True on success, False on any failure (best-effort, never raises).
-    """
-    try:
-        sys.path.insert(0, str(_REPO_ROOT / "scripts" / "lib"))
-        from dispatch_register import append_event
-
-        event_type = str(receipt.get("event_type") or receipt.get("event") or "").lower()
-        status = str(receipt.get("status", "")).lower()
-        gate = str(receipt.get("gate", "")).lower()
-        dispatch_id = str(receipt.get("dispatch_id", ""))
-        terminal = str(receipt.get("terminal", ""))
-        feature_id = str(receipt.get("feature_id", ""))
-        pr_number = receipt.get("pr_number")
-        if pr_number is None:
-            pr_number = receipt.get("metadata", {}).get("pr_number") if isinstance(receipt.get("metadata"), dict) else None
-        try:
-            pr_number = int(pr_number) if pr_number is not None else None
-        except (ValueError, TypeError):
-            pr_number = None
-
-        SUCCESS_STATUSES = {"success", "completed", "complete", "ok", ""}
-        FAILURE_STATUSES = {"failed", "failure", "error", "blocked"}
-
-        register_event = None
-        if event_type in ("task_complete", "task_completed"):
-            if status in FAILURE_STATUSES:
-                register_event = "dispatch_failed"
-            elif status in SUCCESS_STATUSES:
-                register_event = "dispatch_completed"
-            else:
-                return False
-        elif event_type == "task_failed":
-            register_event = "dispatch_failed"
-        elif event_type == "task_timeout":
-            # task_timeout → dispatch_failed: terminal failure semantic.
-            # no_confirmation-style transient timeouts use a different event flow
-            # and don't reach here. Map separately if/when needed (TBD).
-            register_event = "dispatch_failed"
-        elif event_type in ("task_started", "task_start", "dispatch_start"):
-            register_event = "dispatch_started"
-        elif event_type == "review_gate_request":
-            if gate != "codex_gate":
-                return False
-            register_event = "gate_requested"
-        else:
-            return False
-
-        return append_event(
-            register_event,
-            dispatch_id=dispatch_id,
-            pr_number=pr_number,
-            feature_id=feature_id,
-            terminal=terminal,
-            gate=gate,
-        )
-    except Exception:
-        return False
-
-
-def _count_quality_violations(receipt: dict) -> int:
-    """Count violations that _register_quality_open_items WILL create (after dedup).
-
-    Uses the same dedup_key logic as _register_quality_open_items so the
-    pre-computed count matches the actual creation count.
-    Returns 0 when quality_advisory or t0_recommendation is absent.
-    """
-    advisory = receipt.get("quality_advisory") or {}
-    rec = advisory.get("t0_recommendation") or {}
-    open_items = rec.get("open_items") or []
-    if not open_items:
-        return 0
-    # Mirror dedup_key from _register_quality_open_items: qa:{check_id}:{full_path}:{symbol}.
-    seen_keys: set = set()
-    for item in open_items:
-        check_id = str(item.get("check_id", "unknown"))
-        file_path = str(item.get("file", "")) or "unknown"
-        symbol = str(item.get("symbol") or "")
-        seen_keys.add(f"qa:{check_id}:{file_path}:{symbol}")
-    return len(seen_keys)
-
-
-def _count_quality_violations_against_store(receipt: Dict[str, Any]) -> int:
-    """Dry-run dedup count: how many open items _register_quality_open_items WOULD create.
-
-    Mirrors the dedup_key construction of _register_quality_open_items and
-    consults the on-disk open-items store via OpenItemsManager.load_items()
-    so the count reflects items that are NOT already tracked (any status).
-
-    This is read-only: it never writes to open_items.json. It exists so the
-    enrichment step can record an accurate ``open_items_created`` value
-    BEFORE the receipt's idempotency check, without mutating state for
-    receipts that are later skipped as duplicates.
-
-    Returns:
-        Count of unique dedup keys that would result in new open items.
-        0 when quality_advisory or t0_recommendation is absent or on any error.
-    """
-    try:
-        advisory = receipt.get("quality_advisory")
-        if not isinstance(advisory, dict):
-            return 0
-        rec = advisory.get("t0_recommendation")
-        if not isinstance(rec, dict):
-            return 0
-        open_items = rec.get("open_items") or []
-        if not open_items:
-            return 0
-
-        # Load existing dedup keys (any status) from the OI store.
-        existing_keys: set = set()
-        try:
-            oim = _get_open_items_manager()
-            data = oim.load_items()
-            for item in data.get("items", []):
-                key = item.get("dedup_key")
-                if key:
-                    existing_keys.add(key)
-        except Exception as exc:
-            _emit("WARN", "oi_dryrun_load_failed", error=str(exc))
-            # On error, fall back to intra-receipt-only dedup so the count is
-            # still useful (matches pre-fix behaviour for first-run receipts).
-            existing_keys = set()
-
-        seen_keys: set = set()
-        for item in open_items:
-            check_id = str(item.get("check_id", "unknown"))
-            file_path = str(item.get("file", "")) or "unknown"
-            symbol = str(item.get("symbol") or "")
-            dedup_key = f"qa:{check_id}:{file_path}:{symbol}"
-            if dedup_key in existing_keys:
-                continue
-            seen_keys.add(dedup_key)
-        return len(seen_keys)
-    except Exception as exc:
-        _emit("WARN", "oi_dryrun_count_failed", error=str(exc))
-        return 0
-
-
-def _register_quality_open_items(receipt: Dict[str, Any]) -> int:
-    """Best-effort: register quality advisory violations as tracked open items.
-
-    Reads t0_recommendation.open_items[] from the enriched receipt, creates
-    open items with dedup keys, and ALWAYS writes a sidecar summary for the
-    receipt processor to include in T0 notifications (even when clean).
-
-    Returns:
-        Count of newly created open items.
-    """
-    try:
-        advisory = receipt.get("quality_advisory")
-        if not isinstance(advisory, dict):
-            return 0
-
-        rec = advisory.get("t0_recommendation")
-        if not isinstance(rec, dict):
-            return 0
-
-        dispatch_id = str(receipt.get("dispatch_id") or "unknown")
-        report_path = str(receipt.get("report_path") or "")
-        pr_id = str(receipt.get("pr_id") or "")
-
-        new_ids: List[str] = []
-        counts = {"blocker": 0, "warn": 0, "info": 0}
-
-        open_items = rec.get("open_items") or []
-
-        if open_items:
-            oim = _get_open_items_manager()
-
-            for item in open_items:
-                try:
-                    check_id = str(item.get("check_id", "unknown"))
-                    file_path = str(item.get("file", ""))
-                    symbol = str(item.get("symbol") or "")
-                    raw_severity = str(item.get("severity", "info"))
-                    mapped_severity = _SEVERITY_MAP.get(raw_severity, "info")
-                    title = str(item.get("item", ""))
-                    # Build dedup key: qa:{check_id}:{full_path}:{symbol} (OI-1176 fix)
-                    dedup_key = f"qa:{check_id}:{file_path or 'unknown'}:{symbol}"
-
-                    item_id, created = oim.add_item_programmatic(
-                        title=title,
-                        severity=mapped_severity,
-                        dispatch_id=dispatch_id,
-                        report_path=report_path,
-                        pr_id=pr_id,
-                        details=f"file={file_path}, symbol={symbol}" if symbol else f"file={file_path}",
-                        dedup_key=dedup_key,
-                        source="quality_advisory",
-                    )
-
-                    counts[mapped_severity] = counts.get(mapped_severity, 0) + 1
-                    if created:
-                        new_ids.append(item_id)
-                except Exception as exc:
-                    _emit("WARN", "quality_oi_item_failed", error=str(exc))
-
-
-        return len(new_ids)
-
-    except Exception as exc:
-        _emit("WARN", "quality_oi_registration_failed", error=str(exc))
-        return 0
-
-
-def append_receipt_payload(
-    receipt: Dict[str, Any],
-    *,
-    receipts_file: Optional[str] = None,
-    cache_window_seconds: int = 300,
-    skip_enrichment: bool = False,
-) -> AppendResult:
-    if not isinstance(receipt, dict):
-        raise AppendReceiptError("invalid_receipt_type", EXIT_INVALID_INPUT, "Receipt payload must be a JSON object")
-
-    # Enrich completion receipts with quality advisory and terminal snapshot (best-effort).
-    # Enrichment generates quality_advisory; count must run after it sees real items.
-    if not skip_enrichment:
-        receipt = _enrich_completion_receipt(receipt)
-
-    # Count violations from real quality_advisory (generated by enrichment above).
-    # _enrich_completion_receipt sets open_items_created internally; setdefault avoids
-    # double-counting on that path and handles skip_enrichment=True (returns 0 when
-    # quality_advisory is absent).
-    receipt.setdefault("open_items_created", _count_quality_violations(receipt))
-
-    # Route ghost gate receipts (dispatch_id="unknown" + gate event) to gate_events.ndjson.
-    # review_gate_request with empty/missing dispatch_id is redirected here (pre-existing
-    # behaviour from PR #255, see ghost_receipt_filter.py). With PR-2 fix, callers supply
-    # a real dispatch_id so is_ghost_dispatch_id() returns False and receipts land normally
-    # in t0_receipts.ndjson. This is documented intent, not a regression.
-    if receipts_file is None and should_route_to_gate_stream(receipt):
-        try:
-            paths = ensure_env()
-            state_dir = Path(paths["VNX_STATE_DIR"])
-            receipts_file = str(gate_events_file(state_dir))
-            _emit("INFO", "ghost_receipt_rerouted",
-                  gate=str(receipt.get("gate") or ""),
-                  pr_id=str(receipt.get("pr_id") or ""),
-                  destination=receipts_file)
-        except Exception as exc:
-            _emit("WARN", "ghost_receipt_reroute_failed", error=str(exc))
-
-    event_name = _validate_receipt(receipt)
-    idempotency_key = _compute_idempotency_key(receipt, event_name)
-
-    receipt_path = _resolve_receipts_file(receipts_file).expanduser().resolve()
-    receipt_path.parent.mkdir(parents=True, exist_ok=True)
-
-    lock_path = _lock_file_for(receipt_path)
-    cache_path = _cache_file_for(receipt_path)
-
-    min_epoch = time.time() - max(1, int(cache_window_seconds))
-
-    result: Optional[AppendResult] = None
-
-    try:
-        with lock_path.open("a+", encoding="utf-8") as lock_handle:
-            fcntl.flock(lock_handle.fileno(), fcntl.LOCK_EX)
-
-            cache_entries = _load_cache(cache_path, min_epoch)
-            recent_keys = {entry["key"] for entry in cache_entries}
-
-            if idempotency_key in recent_keys:
-                _write_cache(cache_path, cache_entries)
-                result = AppendResult(
-                    status="duplicate",
-                    receipts_file=receipt_path,
-                    idempotency_key=idempotency_key,
-                )
-            else:
-                try:
-                    with receipt_path.open("a", encoding="utf-8") as receipts_handle:
-                        receipts_handle.write(json.dumps(receipt, separators=(",", ":"), sort_keys=False))
-                        receipts_handle.write("\n")
-                except OSError as exc:
-                    raise AppendReceiptError("receipt_write_failed", EXIT_IO_ERROR, f"Failed to append receipt: {exc}") from exc
-
-                cache_entries.append({"ts": time.time(), "key": idempotency_key})
-                _write_cache(cache_path, cache_entries)
-
-                result = AppendResult(
-                    status="appended",
-                    receipts_file=receipt_path,
-                    idempotency_key=idempotency_key,
-                )
-    except AppendReceiptError:
-        raise
-    except OSError as exc:
-        raise AppendReceiptError("lock_failed", EXIT_LOCK_ERROR, f"Failed to acquire append lock: {exc}") from exc
-
-    # Best-effort post-append hooks (skipped for skip_enrichment=True lightweight events).
-    # IMPORTANT: _register_quality_open_items must run only AFTER the idempotency check
-    # has accepted this receipt as 'appended'. A receipt that hits the duplicate cache
-    # must not mutate open_items.json or CQS state. Run it here (post-lock release is
-    # fine — add_item_programmatic uses its own lock file).
-    if result is not None and result.status == "appended" and not skip_enrichment:
-        _register_quality_open_items(receipt)
-
-        _update_confidence_from_receipt(receipt)
-
-        _emit_dispatch_register(receipt)
-
-        _maybe_trigger_state_rebuild(receipt)
-
-        _trigger_receipt_classifier(receipt)
-
-    return result
-
-
-def _update_confidence_from_receipt(receipt: Dict[str, Any]) -> None:
-    """Wire dispatch outcome into pattern confidence scores (best-effort)."""
-    try:
-        SUCCESS_STATUSES = {"success", "completed", "complete", "ok", ""}
-        FAILURE_STATUSES = {"failed", "failure", "error", "blocked"}
-
-        event_type = str(receipt.get("event_type") or receipt.get("event") or "").lower()
-        status = str(receipt.get("status", "")).lower()
-
-        if event_type in ("task_complete", "task_completed"):
-            if status in FAILURE_STATUSES:
-                outcome = "failure"
-            elif status in SUCCESS_STATUSES:
-                outcome = "success"
-            else:
-                return  # unknown status — don't update confidence
-        elif event_type == "task_failed":
-            outcome = "failure"
-        else:
-            return
-
-        dispatch_id = str(receipt.get("dispatch_id") or "")
-        terminal = str(receipt.get("terminal") or "")
-        if not dispatch_id:
-            return
-
-        state_dir = resolve_state_dir(__file__)
-
-        db_path = state_dir / "quality_intelligence.db"
-        if not db_path.exists():
-            return
-
-        sys.path.insert(0, str(SCRIPT_DIR / "lib"))
-        from intelligence_persist import update_confidence_from_outcome
-        update_confidence_from_outcome(db_path, dispatch_id, terminal, outcome)
-    except Exception as exc:
-        _emit("WARN", "confidence_update_failed", error=str(exc))
-
-
-def _trigger_receipt_classifier(receipt: Dict[str, Any]) -> None:
-    """Best-effort fire of the adaptive receipt classifier (ARC-3).
-
-    Disabled by default; opt-in via VNX_RECEIPT_CLASSIFIER_ENABLED=1. Never
-    raises — the receipt writer must remain on its happy path even if the
-    classifier import or subprocess spawn fails.
-    """
-    if os.environ.get("VNX_RECEIPT_CLASSIFIER_ENABLED", "0") != "1":
-        return
-    try:
-        sys.path.insert(0, str(SCRIPT_DIR / "lib"))
-        from receipt_classifier import trigger_receipt_classifier_async
-        action = trigger_receipt_classifier_async(receipt)
-        if action:
-            _emit("INFO", "receipt_classifier_action", action=action)
-    except Exception as exc:
-        _emit("WARN", "receipt_classifier_trigger_failed", error=str(exc))
-
-
-def _maybe_trigger_state_rebuild(receipt: Dict[str, Any]) -> None:
-    """Trigger state rebuild via shared throttled helper. Best-effort."""
-    event_type = str(receipt.get("event_type") or receipt.get("event") or "").lower()
-
-    TRIGGER_EVENTS = {
-        "task_complete", "task_completed", "completion", "complete",
-        "task_failed", "task_timeout",
-        "dispatch_promoted", "dispatch_started",
-    }
-    if event_type not in TRIGGER_EVENTS:
-        return
-
-    try:
-        sys.path.insert(0, str(_REPO_ROOT / "scripts" / "lib"))
-        from state_rebuild_trigger import maybe_trigger_state_rebuild
-        maybe_trigger_state_rebuild(event_type=event_type)
-    except Exception:
-        pass  # best-effort
+# Register this module as the active facade so submodules can resolve
+# patchable names back to whichever module the test loaded.
+register_facade(sys.modules[__name__])
 
 
 def _parse_input(receipt_json: Optional[str], receipt_file: Optional[str]) -> Dict[str, Any]:
