+++ scripts/lib/append_receipt_internals/payload.py
@@ -1,0 +1,168 @@
+"""append_receipt_payload pipeline + post-append hooks."""
+
+from __future__ import annotations
+
+import os
+import sys
+from pathlib import Path
+from typing import Any, Dict, Optional
+
+from .common import (
+    AppendReceiptError,
+    AppendResult,
+    EXIT_INVALID_INPUT,
+    REPO_ROOT,
+    SCRIPTS_DIR,
+    _emit,
+    facade,
+)
+from .idempotency import (
+    _compute_idempotency_key,
+    _cache_file_for,
+    _resolve_receipts_file,
+    _write_receipt_under_lock,
+)
+from .validation import _validate_receipt
+
+
+def _maybe_reroute_to_gate_stream(receipt: Dict[str, Any], receipts_file: Optional[str]) -> Optional[str]:
+    """Route ghost gate receipts (dispatch_id="unknown" + gate event) to gate_events.ndjson."""
+    if receipts_file is not None or not facade.should_route_to_gate_stream(receipt):
+        return receipts_file
+    try:
+        paths = facade.ensure_env()
+        state_dir = Path(paths["VNX_STATE_DIR"])
+        rerouted = str(facade.gate_events_file(state_dir))
+        _emit("INFO", "ghost_receipt_rerouted",
+              gate=str(receipt.get("gate") or ""),
+              pr_id=str(receipt.get("pr_id") or ""),
+              destination=rerouted)
+        return rerouted
+    except Exception as exc:
+        _emit("WARN", "ghost_receipt_reroute_failed", error=str(exc))
+        return receipts_file
+
+
+def _run_post_append_hooks(receipt: Dict[str, Any]) -> None:
+    """Best-effort hooks fired after a receipt is successfully appended."""
+    facade._register_quality_open_items(receipt)
+    facade._update_confidence_from_receipt(receipt)
+    facade._emit_dispatch_register(receipt)
+    facade._maybe_trigger_state_rebuild(receipt)
+    facade._trigger_receipt_classifier(receipt)
+
+
+def append_receipt_payload(
+    receipt: Dict[str, Any],
+    *,
+    receipts_file: Optional[str] = None,
+    cache_window_seconds: int = 300,
+    skip_enrichment: bool = False,
+) -> AppendResult:
+    if not isinstance(receipt, dict):
+        raise AppendReceiptError("invalid_receipt_type", EXIT_INVALID_INPUT, "Receipt payload must be a JSON object")
+
+    if not skip_enrichment:
+        receipt = facade._enrich_completion_receipt(receipt)
+
+    receipt.setdefault("open_items_created", facade._count_quality_violations(receipt))
+
+    receipts_file = _maybe_reroute_to_gate_stream(receipt, receipts_file)
+
+    event_name = _validate_receipt(receipt)
+    idempotency_key = _compute_idempotency_key(receipt, event_name)
+
+    receipt_path = _resolve_receipts_file(receipts_file).expanduser().resolve()
+    receipt_path.parent.mkdir(parents=True, exist_ok=True)
+    cache_path = _cache_file_for(receipt_path)
+
+    result = _write_receipt_under_lock(
+        receipt,
+        receipt_path,
+        cache_path,
+        idempotency_key,
+        cache_window_seconds,
+    )
+
+    if result.status == "appended" and not skip_enrichment:
+        _run_post_append_hooks(receipt)
+
+    return result
+
+
+def _update_confidence_from_receipt(receipt: Dict[str, Any]) -> None:
+    """Wire dispatch outcome into pattern confidence scores (best-effort)."""
+    try:
+        SUCCESS_STATUSES = {"success", "completed", "complete", "ok", ""}
+        FAILURE_STATUSES = {"failed", "failure", "error", "blocked"}
+
+        event_type = str(receipt.get("event_type") or receipt.get("event") or "").lower()
+        status = str(receipt.get("status", "")).lower()
+
+        if event_type in ("task_complete", "task_completed"):
+            if status in FAILURE_STATUSES:
+                outcome = "failure"
+            elif status in SUCCESS_STATUSES:
+                outcome = "success"
+            else:
+                return
+        elif event_type == "task_failed":
+            outcome = "failure"
+        else:
+            return
+
+        dispatch_id = str(receipt.get("dispatch_id") or "")
+        terminal = str(receipt.get("terminal") or "")
+        if not dispatch_id:
+            return
+
+        state_dir = facade.resolve_state_dir(__file__)
+
+        db_path = state_dir / "quality_intelligence.db"
+        if not db_path.exists():
+            return
+
+        sys.path.insert(0, str(SCRIPTS_DIR / "lib"))
+        from intelligence_persist import update_confidence_from_outcome
+        update_confidence_from_outcome(db_path, dispatch_id, terminal, outcome)
+    except Exception as exc:
+        _emit("WARN", "confidence_update_failed", error=str(exc))
+
+
+def _trigger_receipt_classifier(receipt: Dict[str, Any]) -> None:
+    """Best-effort fire of the adaptive receipt classifier (ARC-3).
+
+    Disabled by default; opt-in via VNX_RECEIPT_CLASSIFIER_ENABLED=1. Never
+    raises — the receipt writer must remain on its happy path even if the
+    classifier import or subprocess spawn fails.
+    """
+    if os.environ.get("VNX_RECEIPT_CLASSIFIER_ENABLED", "0") != "1":
+        return
+    try:
+        sys.path.insert(0, str(SCRIPTS_DIR / "lib"))
+        from receipt_classifier import trigger_receipt_classifier_async
+        action = trigger_receipt_classifier_async(receipt)
+        if action:
+            _emit("INFO", "receipt_classifier_action", action=action)
+    except Exception as exc:
+        _emit("WARN", "receipt_classifier_trigger_failed", error=str(exc))
+
+
+def _maybe_trigger_state_rebuild(receipt: Dict[str, Any]) -> None:
+    """Trigger state rebuild via shared throttled helper. Best-effort."""
+    event_type = str(receipt.get("event_type") or receipt.get("event") or "").lower()
+
+    TRIGGER_EVENTS = {
+        "task_complete", "task_completed", "completion", "complete",
+        "task_failed", "task_timeout",
+        "dispatch_promoted", "dispatch_started",
+    }
+    if event_type not in TRIGGER_EVENTS:
+        return
+
+    try:
+        sys.path.insert(0, str(REPO_ROOT / "scripts" / "lib"))
+        from state_rebuild_trigger import maybe_trigger_state_rebuild
+        maybe_trigger_state_rebuild(event_type=event_type)
+    except Exception:
+        pass
